Move some function to pipeline package
Move runsources to pipeline package

Move RunConditions to pipeline package

Move RunTargets to pipeline package
