+++ pkg/core/pipeline/sort.go
@@ -1,4 +1,4 @@
-package engine
+package pipeline
 
 import (
 	"errors"
